Modified location of cache directory for use with Leeds HPC cluster (Arc3) and some PEP-8 refactoring.
Changed the cache directory for UKCensusAPI, household_microsynth, and microsimulation packages to suit implementation on HPC (Arc3). Previously, the UKCensusAPI package specified a cache directory in '/tmp/UKCensusAPI', which is not ideal when located on HPC platform with multiple users. The 3 separate places now point to a single global cache directory at the top level of the project. This means the NOMIS_API_KEY file can be included only once (environment variables are difficult to manually specify on Arc3) in this cache. This is easier for initial setup, more secure (keeping API key out of public /tmp/ directory), and also easier for new users to understand.
Lot's of files in microsimulation package have also been refactored to keep to PEP-8 style guidelines.'